+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - fixed a crash that occured when all todo items were removed
 - added a menu entry to show the changelog in the help menu
 - fixed a design glitch with the visibility of the todo reminder date selector
+- now a warning is issued if you want to open the todo lists, but haven't configured them yet
 
 ## 0.71
 - you are now able to print notes (text or markdown)